SpotTask tested with latest master branch code changes.
