+++ python/lsst/eotest/sensor/spotTask.py
@@ -18,16 +18,58 @@ import lsst.meas.extensions.shapeHSM
 from .MaskedCCD import MaskedCCD
 from .AmplifierGeometry import parse_geom_kwd
 
+def make_ccd_mosaic(infile, bias_frame=None, gains=None, fit_order=1):
+    """Combine amplifier image arrays into a single calibrated CCD image mosaic."""
+    ccd = MaskedCCD(infile, bias_frame=bias_frame)
+    foo = fits.open(infile)
+    datasec = parse_geom_kwd(foo[1].header['DATASEC'])
+    nx_segments = 8
+    ny_segments = 2
+    nx = nx_segments*(datasec['xmax'] - datasec['xmin'] + 1)
+    ny = ny_segments*(datasec['ymax'] - datasec['ymin'] + 1)
+    mosaic = np.zeros((ny, nx), dtype=np.float32)
+
+    for ypos in range(ny_segments):
+        for xpos in range(nx_segments):
+            amp = ypos*nx_segments + xpos + 1
+
+            detsec = parse_geom_kwd(foo[amp].header['DETSEC'])
+            xmin = nx - max(detsec['xmin'], detsec['xmax'])
+            xmax = nx - min(detsec['xmin'], detsec['xmax']) + 1
+            ymin = ny - max(detsec['ymin'], detsec['ymax'])
+            ymax = ny - min(detsec['ymin'], detsec['ymax']) + 1
+            #
+            # Extract bias-subtracted image for this segment
+            #
+            segment_image = ccd.unbiased_and_trimmed_image(amp, fit_order=fit_order)
+            subarr = segment_image.getImage().getArray()
+            #
+            # Determine flips in x- and y- direction
+            #
+            if detsec['xmax'] > detsec['xmin']: # flip in x-direction
+                subarr = subarr[:, ::-1]
+            if detsec['ymax'] > detsec['ymin']: # flip in y-direction
+                subarr = subarr[::-1, :]
+            #
+            # Convert from ADU to e-
+            #
+            if gains is not None:
+                subarr *= gains[amp]
+            #
+            # Set sub-array to the image mosaic
+            #
+            mosaic[ymin:ymax, xmin:xmax] = subarr
+
+    image = afwImage.ImageF(mosaic)
+    return image
+
 class SpotConfig(pexConfig.Config):
     """Configuration for Spot analysis task"""
     minpixels = pexConfig.Field("Minimum number of pixels above detection threshold", 
                                 int, default=10)
     nsig = pexConfig.Field("Source footprint threshold in number of standard deviations.", 
                            float, default=10)
-    temp_set_point = pexConfig.Field("Required temperature (C) set point",
-                                     float, default=-95.)
-    temp_set_point_tol = pexConfig.Field("Required temperature set point tolerance (degrees C)",
-                                         float, default=1.)
+    bgbinsize = pexConfig.Field("Bin size for background estimation.", int, default=10)
     output_dir = pexConfig.Field("Output directory", str, default='.')
     output_file = pexConfig.Field("Output filename", str, default=None)
     verbose = pexConfig.Field("Turn verbosity on", bool, default=True)
@@ -40,32 +82,35 @@ class SpotTask(pipeBase.Task):
 
     @pipeBase.timeMethod
     def run(self, sensor_id, infile, gains, bias_frame=None):
-
-        if self.config.verbose:
-            self.log.info("Input file:")
-            self.log.info("  {0}".format(infile))
         #
         # Process a CCD image mosaic
         #
         if self.config.verbose:
             self.log.info("processing {0}".format(infile))
-        image = imutils.make_ccd_mosaic(infile, bias_frame=bias_frame, gains=gains)
+        image = make_ccd_mosaic(infile, bias_frame=bias_frame, gains=gains)
         exposure = afwImage.ExposureF(image.getBBox())
         exposure.setImage(image)
         #
         # Set up characterize task configuration
         #
         nsig = self.config.nsig
+        bgbinsize = self.config.bgbinsize
         minpixels = self.config.minpixels
         charConfig = CharacterizeImageConfig()
         charConfig.doMeasurePsf = False
         charConfig.doApCorr = False
         charConfig.repair.doCosmicRay = False
         charConfig.detection.minPixels = minpixels
-        charConfig.detection.background.binSize = 10
+        charConfig.detection.background.binSize = bgbinsize
         charConfig.detection.thresholdType = "stdev"
         charConfig.detection.thresholdValue = nsig
-        charConfig.measurement.plugins.names |= ["ext_shapeHSM_HsmSourceMoments"]
+        hsm_plugins = set(["ext_shapeHSM_HsmShapeBj",
+                           "ext_shapeHSM_HsmShapeLinear",
+                           "ext_shapeHSM_HsmShapeKsb",
+                           "ext_shapeHSM_HsmShapeRegauss",
+                           "ext_shapeHSM_HsmSourceMoments",
+                           "ext_shapeHSM_HsmPsfMoments"])  
+        charConfig.measurement.plugins.names |= hsm_plugins
         charTask = CharacterizeImageTask(config=charConfig)
         result = charTask.characterize(exposure)
         src = result.sourceCat
